+++ Assets/Scripts/LocalPlayerBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0085b50e5a4c6c343b1f4bdab30bf1a3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
